[REFACTOR] 필터 컴포넌트 계층 변경
