+++ components/SupportPrograms/SupportProgramFilters/FilterDetail/index.ts
@@ -1,0 +1,1 @@
+export { default as FilterDetail } from './FilterDetail';
