+++ components/SupportPrograms/SupportProgramFilters/FilterTable/index.ts
@@ -1,0 +1,2 @@
+export { default as FilterTable } from './FilterTable';
+export { default as FilterTableRow } from './FilterTableRow';
