+++ components/SupportPrograms/SupportProgramFilters/TypeFilters/index.ts
@@ -1,0 +1,1 @@
+export { default as TypeFilters } from './TypeFilters';
