+++ components/SupportPrograms/SupportProgramFilters/index.ts
@@ -1,2 +1,6 @@
-export { default as SupportProgramFilters } from './SupportProgramFilters';
-export { default as useSupportProgramFilters } from './SupportProgramFilters.hooks';
+export * from './SupportProgramFilters.hooks';
+export * from './SupportProgramFilters.types';
+export * from './TypeFilters';
+export * from './FilterDetail';
+export * from './HostSearch';
+export * from './FilterTable';
