+++ components/SupportPrograms/SupportPrograms.tsx
@@ -1,78 +1,113 @@
-import TypeFilters from './SupportProgramFilters/TypeFilters';
-
-import { useSupportProgramFilters, SupportProgramFilters } from './SupportProgramFilters';
-import MultipleSelectionFilters from './SupportProgramFilters/MultipleSelectionFilters';
-
 import {
+  useSupportProgramFilters,
+  TypeFilters,
   useClientFilter,
-  FilterOptionsQuery,
-} from './SupportProgramFilters/SupportProgramFilters.hooks';
-import { AreaEnum, SupportProgramTypeEnum, TargetCompanyAgeEnum } from '../../graphql';
+  FilterTable,
+  FilterTableRow,
+  FilterDetail,
+} from './SupportProgramFilters';
+
+import { useModal } from '../../commonUi/Modal';
+
 import { TARGET_COMPANY_AGE_TEXTS, AREA_TEXTS } from '../../constants/supportPrograms';
+
+import type { Area, TargetCompanyAge, Host, Type } from './SupportProgramFilters';
 
 function SupportPrograms() {
   const filterQuery = useSupportProgramFilters();
 
-  const [activeTypes, toggleTypes, filteredActiveTypes] = useClientFilter<SupportProgramTypeEnum>({
+  const [activeTypes, toggleTypes, filteredActiveTypes] = useClientFilter<Type>({
     multiple: false,
   });
 
-  const [activeAges, toggleAges, filteredActiveAges] = useClientFilter<TargetCompanyAgeEnum>({
+  const [activeAges, toggleAges, filteredActiveAges] = useClientFilter<TargetCompanyAge>({
     multiple: true,
   });
 
-  const [activeAreas, toggleAreas, filteredActiveAreas] = useClientFilter<AreaEnum>({
+  const [activeAreas, toggleAreas, filteredActiveAreas] = useClientFilter<Area>({
     multiple: true,
   });
 
-  const [activeHosts, toggleHosts, filteredActiveHosts] = useClientFilter<
-    FilterOptionsQuery['filterOptions']['hosts'][number]
-  >({
+  const [activeHosts, toggleHosts, filteredActiveHosts] = useClientFilter<Host>({
     multiple: true,
   });
+  const { hide } = useModal();
 
   return (
     <div>
       {filterQuery.isSuccess && (
-        <SupportProgramFilters
-          top={
-            <TypeFilters
-              allTypes={filterQuery.data.types}
-              activeTypes={activeTypes}
-              onClick={toggleTypes}
-            />
-          }
-          ages={
-            <MultipleSelectionFilters
-              title="창업 기간"
-              toggle={toggleAges}
-              keyExtractor={(data) => data}
-              data={filterQuery.data.targetCompanyAges}
-              activeData={activeAges}
-              renderItemText={(data) => TARGET_COMPANY_AGE_TEXTS[data]}
-            />
-          }
-          areas={
-            <MultipleSelectionFilters
-              title="지원 분야"
-              toggle={toggleAreas}
-              keyExtractor={(data) => data}
-              data={filterQuery.data.areas}
-              activeData={activeAreas}
-              renderItemText={(data) => AREA_TEXTS[data]}
-            />
-          }
-          hosts={
-            <MultipleSelectionFilters
-              title="주관 기관"
-              toggle={toggleHosts}
-              keyExtractor={(data) => data.id}
-              data={filterQuery.data.hosts}
-              activeData={activeHosts}
-              renderItemText={(data) => data.meta.name}
-            />
-          }
-        />
+        <>
+          <TypeFilters
+            allTypes={filterQuery.data.types}
+            activeTypes={activeTypes}
+            onClick={toggleTypes}
+          />
+          <FilterTable
+            ages={
+              <FilterTableRow
+                title="창업 기간"
+                toggle={toggleAges}
+                keyExtractor={(data) => data}
+                data={filterQuery.data.targetCompanyAges}
+                activeData={activeAges}
+                renderItemText={(data) => TARGET_COMPANY_AGE_TEXTS[data]}
+                Detail={
+                  <FilterDetail
+                    title="창업 기간"
+                    toggle={toggleAges}
+                    keyExtractor={(data) => data}
+                    data={filterQuery.data.targetCompanyAges}
+                    activeData={activeAges}
+                    renderItemText={(data) => TARGET_COMPANY_AGE_TEXTS[data]}
+                    onClose={hide}
+                  />
+                }
+              />
+            }
+            areas={
+              <FilterTableRow
+                title="지원 분야"
+                toggle={toggleAreas}
+                keyExtractor={(data) => data}
+                data={filterQuery.data.areas}
+                activeData={activeAreas}
+                renderItemText={(data) => AREA_TEXTS[data]}
+                Detail={
+                  <FilterDetail
+                    title="지원 분야"
+                    toggle={toggleAreas}
+                    keyExtractor={(data) => data}
+                    data={filterQuery.data.areas}
+                    activeData={activeAreas}
+                    renderItemText={(data) => AREA_TEXTS[data]}
+                    onClose={hide}
+                  />
+                }
+              />
+            }
+            hosts={
+              <FilterTableRow
+                title="주관 기관"
+                toggle={toggleHosts}
+                keyExtractor={(data) => data.id}
+                data={filterQuery.data.hosts}
+                activeData={activeHosts}
+                renderItemText={(data) => data.meta.name}
+                Detail={
+                  <FilterDetail
+                    title="주관 기관"
+                    toggle={toggleHosts}
+                    keyExtractor={(data) => data.id}
+                    data={filterQuery.data.hosts}
+                    activeData={activeHosts}
+                    renderItemText={(data) => data.meta.name}
+                    onClose={hide}
+                  />
+                }
+              />
+            }
+          />
+        </>
       )}
     </div>
   );
